globalcrm/requirements.txt
asgiref==3.6.0
black==23.11.0
click==8.1.7
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
django-multiupload==0.6.1
django-scheduler==0.10.1
gunicorn==21.2.0
icalendar==5.0.11
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
prettier==0.0.7
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
